requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
deprecation==2.1.0
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
google-search-results==2.4.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
fpdf2>=2.7.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.7.0
numpy==2.3.3
openai==2.2.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
postgrest==2.21.1
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
realtime==2.21.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
storage3==2.21.1
streamlit==1.50.0
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
playwright==1.42.0
fpdf2==2.7.9
feedparser==6.0.11
beautifulsoup4==4.14.2



